+++ rotkehlchen/db/cache.py
@@ -26,6 +26,7 @@ class DBCacheStatic(Enum):
     LAST_CREATE_REMINDER_CHECK_TS: Final = 'last_create_reminder_check_ts'
     LAST_GRAPH_DELEGATIONS_CHECK_TS: Final = 'last_graph_delegations_check_ts'
     LAST_GNOSISPAY_QUERY_TS: Final = 'last_gnosispay_query_ts'
+    LAST_SPARK_ASSETS_UPDATE: Final = 'last_spark_assets_update'
 
 
 class LabeledLocationArgsType(TypedDict):